initial checkin to liberty branch

+++ deployment_scripts/puppet/modules/l23network/lib/puppet/provider/l2_port/lnx.rb
@@ -1,10 +1,8 @@
 require File.join(File.dirname(__FILE__), '..','..','..','puppet/provider/lnx_base')
 
 Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base) do
-  defaultfor :osfamily    => :linux
-  commands   :iproute     => 'ip',
-             :ethtool_cmd => 'ethtool',
-             :brctl       => 'brctl',
+  defaultfor :kernel    => :linux
+  commands   :ethtool_cmd => 'ethtool',
              :pkill       => 'pkill'
 
 
@@ -47,7 +45,7 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
     @property_flush = {}.merge! @resource
     # todo: divide simple creating interface and vlan
     begin
-      iproute('link', 'add', 'link', @resource[:vlan_dev], 'name', @resource[:interface], 'type', 'vlan', 'id', @resource[:vlan_id])
+      self.class.iproute(['link', 'add', 'link', @resource[:vlan_dev], 'name', @resource[:interface], 'type', 'vlan', 'id', @resource[:vlan_id]])
     rescue
       # Some time interface may be created by OS init scripts. It's a normal for Ubuntu.
       raise if ! self.class.iface_exist? @resource[:interface]
@@ -59,7 +57,7 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
     debug("DESTROY resource: #{@resource}")
     # todo: Destroing of L2 resource -- is a putting interface to the DOWN state.
     #       Or remove, if ove a vlan interface
-    #iproute('--force', 'addr', 'flush', 'dev', @resource[:interface])
+    #iproute(['--force', 'addr', 'flush', 'dev', @resource[:interface]])
   end
 
   def flush
@@ -73,23 +71,35 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
       end
       #
       # FLUSH changed properties
-      if @property_flush.has_key? :bond_master
+      if @property_flush.has_key? :bond_master and @property_flush[:bond_master] != :absent
         bond = @old_property_hash[:bond_master]
+        if self.class.ipaddr_exist? @resource[:interface]
+          # remove all IP addresses from member of bond. This should be done on device in UP state
+          self.class.addr_flush(@resource[:interface])
+        end
         # putting interface to the down-state, because add/remove upped interface impossible. undocumented kern.behavior.
-        iproute('--force', 'link', 'set', 'dev', @resource[:interface], 'down')
-        if bond and bond != :absent
+        self.class.interface_down(@resource[:interface], true)
+        if bond and bond != :absent and File.exist?("/sys/class/net/#{@resource[:interface]}/master/bonding/slaves")
           # remove interface from bond, if one included to it
           debug("Remove interface '#{@resource[:interface]}' from bond '#{bond}'.")
           File.open("/sys/class/net/#{@resource[:interface]}/master/bonding/slaves", "a") {|f| f << "-#{@resource[:interface]}"}
         end
         if ! @property_flush[:bond_master].nil? and @property_flush[:bond_master] != :absent
           # add interface as slave to bond
-          debug("Add interface '#{@resource[:interface]}' to bond '#{@property_flush[:bond_master]}'.")
-          File.open("/sys/class/net/#{@property_flush[:bond_master]}/bonding/slaves", "a") {|f| f << "+#{@resource[:interface]}"}
+          if File.exist? "/sys/class/net/#{@property_flush[:bond_master]}/bonding/slaves"
+            # If port is bond member and bond still doesn't exist we skip the adding to bond action
+            # Bond will do it by itself during bond creation
+            debug("Add interface '#{@resource[:interface]}' to bond '#{@property_flush[:bond_master]}'.")
+            File.open("/sys/class/net/#{@property_flush[:bond_master]}/bonding/slaves", "a") {|f| f << "+#{@resource[:interface]}"}
+          end
         else
           # port no more member of any bonds
           @property_flush[:port_type] = nil
         end
+        # Up parent interface if this is vlan port
+        self.class.interface_up(@resource[:vlan_dev]) if @resource[:vlan_dev]
+        # Up port
+        self.class.interface_up(@resource[:interface])
       end
       if @property_flush.has_key? :bridge
         # get actual bridge-list. We should do it here,
@@ -100,9 +110,9 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
         debug("Actual-port-bridge-mapping: '#{port_bridges_hash}'")       # it should removed from LNX
         #
         #Flush ipaddr and routes for interface, thah adding to the bridge
-        iproute('route', 'flush', 'dev', @resource[:interface])
-        iproute('addr', 'flush', 'dev', @resource[:interface])
-        iproute('--force', 'link', 'set', 'dev', @resource[:interface], 'down')
+        self.class.route_flush(@resource[:interface])
+        self.class.addr_flush(@resource[:interface])
+        self.class.interface_down(@resource[:interface], true)
         # remove interface from old bridge
         if ! port_bridges_hash[@resource[:interface]].nil?
           br_name = port_bridges_hash[@resource[:interface]][:bridge]
@@ -113,7 +123,7 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
             when :ovs
               self.class.ovs_vsctl(['del-port', br_name, @resource[:interface]])
             when :lnx
-              brctl('delif', br_name, @resource[:interface])
+              self.class.brctl(['delif', br_name, @resource[:interface]])
             else
               #pass
             end
@@ -126,7 +136,7 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
             self.class.ovs_vsctl(['add-port', @property_flush[:bridge], @resource[:interface]])
           when :lnx
             begin
-              brctl('addif', @property_flush[:bridge], @resource[:interface])
+              self.class.brctl(['addif', @property_flush[:bridge], @resource[:interface]])
             rescue
               # Sometimes interface may be automatically added to bridge if config file exists before interface creation,
               # especially vlan interfaces. It appears on CentOS.
@@ -137,7 +147,7 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
             #pass
           end
         end
-        iproute('link', 'set', 'dev', @resource[:interface], 'up')
+        self.class.interface_up(@resource[:interface])
         debug("Change bridge")
       end
       if @property_flush.has_key? :ethtool and @property_flush[:ethtool].is_a? Hash
@@ -163,7 +173,10 @@ Puppet::Type.type(:l2_port).provide(:lnx, :parent => Puppet::Provider::Lnx_base)
       if ! @property_flush[:onboot].nil?
         # Should be after bond, because interface may auto-upped while added to the bond
         debug("Setup UP state for interface '#{@resource[:interface]}'.")
-        iproute('link', 'set', 'dev', @resource[:interface], 'up')
+        # Up parent interface if this is vlan port
+        self.class.interface_up(@resource[:vlan_dev]) if @resource[:vlan_dev]
+        # Up port
+        self.class.interface_up(@resource[:interface])
       end
       if !['', 'absent'].include? @property_flush[:mtu].to_s
         self.class.set_mtu(@resource[:interface], @property_flush[:mtu])
